refactor(analytics, fiam): revert ktx imports
